requirements.txt
Flask>=2.3.4
flask-cors>=3.0.10
google-generativeai>=0.8.0
SpeechRecognition>=3.9.0
pyaudio>=0.2.13
langdetect>=1.0.9
opencv-python>=4.7.0.72
pyautogui>=0.9.53
numpy>=1.26.0
